Trying to get timeouts working in Informix plus fixed Tickets #48-49

+++ src/main/java/nz/co/gregs/dbvolution/datatypes/DBInteger.java
@@ -665,7 +665,7 @@ public class DBInteger extends QueryableDatatype implements NumberResult {
 		if (newLiteralValue == null) {
 			super.setLiteralValue(null);
 		} else {
-			super.setLiteralValue(newLiteralValue);
+			super.setLiteralValue(newLiteralValue.longValue());
 		}
 	}
 
@@ -691,7 +691,7 @@ public class DBInteger extends QueryableDatatype implements NumberResult {
 		if (newLiteralValue == null) {
 			super.setLiteralValue(null);
 		} else {
-			super.setLiteralValue(newLiteralValue);
+			super.setLiteralValue(newLiteralValue.longValue());
 		}
 	}
